+++ docs/dev-demos/src/specs/large-line-group-encode.ts
@@ -161,6 +161,6 @@ export const callback = (chartInstance: IView) => {
 
     chartInstance.getDataById('table').update(newValues);
 
-    chartInstance.runAsync();
+    chartInstance.run();
   });
 };