+++ docs/dev-demos/src/specs/pie-animation.ts
@@ -142,6 +142,6 @@ export const callback = (chartInstance: any) => {
   });
   disappearButton.addEventListener('click', () => {
     chartInstance.getSignalById('animationState').value('disappear');
-    chartInstance.runAsync();
+    chartInstance.run();
   });
 };