+++ packages/vgrammar-wordcloud-shape/src/segmentation.ts
@@ -1,14 +1,20 @@
 import type { CloudWordType, SegmentationInputType, SegmentationOutputType } from './interface';
 import { loadImage } from './util';
 
-export async function loadAndHandleImage(segmentationInput: SegmentationInputType): Promise<CanvasImageSource> {
-  const shapeImage = (await loadImage(segmentationInput.shapeUrl)) as CanvasImageSource;
+export function loadAndHandleImage(segmentationInput: SegmentationInputType): Promise<CanvasImageSource> {
+  const imagePromise = loadImage(segmentationInput.shapeUrl);
 
-  if (segmentationInput.removeWhiteBorder && shapeImage) {
-    return removeBorder(shapeImage, segmentationInput.tempCanvas, segmentationInput.tempCtx);
+  if (!imagePromise) {
+    return null;
   }
 
-  return shapeImage;
+  return imagePromise.then((shapeImage: unknown) => {
+    if (segmentationInput && segmentationInput.removeWhiteBorder && shapeImage) {
+      return removeBorder(shapeImage, segmentationInput.tempCanvas, segmentationInput.tempCtx);
+    }
+
+    return shapeImage;
+  });
 }
 
 /**